+++ lib/socket/socketserver.py
@@ -132,8 +132,8 @@ class SocketServer:
             logging.info(
                 f'New listener with id: {client_id}')
         elif client_data.get('client_type') == 'worker':
-            supported_models = client_data.get('models')
-            if supported_models is None or len(supported_models) < 1:
+            supported_modules = client_data.get('modules')
+            if supported_modules is None or len(supported_modules) < 1:
                 logging.warning(
                     f"New worker connected with id : {client_id}"
                     ", but has no installed modules")